Немного оптимизирована скорость загрузки данных при перегрузке метода transform в дата-провайдере
